+++ service/src/index.ts
@@ -1,3 +1,6 @@
 import { startService } from "./start_service";
 
 await startService({ nodeName: "arch" });
+export { logProfile } from "./cli/utils";
+export { objectDashToCamelCase } from "./cli/utils";
+export { dashToCamelCase } from "./cli/utils";
